+++ R/rMEA_IO.R
@@ -19,8 +19,9 @@
 #'   filenames and correctly label the cases. The strings can be abbreviated.
 #'   If the filenames contains other data the character "x" can be used to skip a position.
 #'   If NA, no attempt to identify cases will be done.
-#' @param idSep If idOrder is not NA, this character will be used as separator between "id", "session", and "group"
-#' information in the filenames.
+#' @param idSep character vector (or object which can be coerced to such) containing regular expression(s).
+#'   If idOrder is not NA, this will be used as separator to split the filenameas and identify "id", "session", and "group"
+#'   informations.
 #' @param removeShortFiles Either NULL or an number ranging from 0 to 1.
 #'  Specifies the proportion of the average file length below which a file should be excluded.
 #'  (E.g. a value of 0.7 will exclude all files with a duration smaller than 70\% of the mean duration of all other files in the directory.)