+++ NEWS.md
@@ -3,6 +3,8 @@
 
 * fixed a bug in plot.MEA when interpolation results in NA vals
 * updated documentation in readMEA
+* all functions accepting _MEAlist_ objects now should accept a list instead and try to parse as _MEAlist_ internally
+
 
 ### v1.0.0.9010
 * more informative error messages